trend_pipeline/requirements.txt
anthropic>=0.40.0
requests>=2.31.0
beautifulsoup4>=4.12.0
lxml>=5.1.0
pytrends>=4.9.2
python-dotenv>=1.0.0
